// data/nepse/types.ts
export type NepseDecodeFunction = (
  var0: number,
  var1: number,
  var2: number,
  var3: number,
  var4: number,
) => number;

export type NepseExports = {
  cdx: NepseDecodeFunction;
  rdx: NepseDecodeFunction;
  bdx: NepseDecodeFunction;
  ndx: NepseDecodeFunction;
  mdx: NepseDecodeFunction;
};